// apps/web/src/api/generated/models/tmdbMovieDetails.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Verdict API
 * OpenAPI documentation for the Verdict backend.
 * OpenAPI spec version: 1.0.0
 */
import type { TmdbGenre } from './tmdbGenre';

export interface TmdbMovieDetails {
  adult: boolean;
  /** @nullable */
  backdrop_path: string | null;
  genres: TmdbGenre[];
  id: number;
  /** @nullable */
  imdb_id: string | null;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  /** @nullable */
  poster_path: string | null;
  release_date: string;
  /** @nullable */
  runtime: number | null;
  tagline: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}